@@ -1,7 +1,8 @@
 {
   "name": "ywc15-api",
   "version": "1.0.0",
-  "description": "RESTFul API Server for YWC#15 Website. Code reuse from YWC#14",
+  "description":
+    "RESTFul API Server for YWC#15 Website. Code reuse from YWC#14",
   "main": "build/bin/server.js",
   "scripts": {
     "eslint": "eslint --quiet **/**/*.js **/*.js **.js",
@@ -31,7 +32,6 @@
   "homepage": "https://github.com/komcal/ywc-register-backend-api#readme",
   "dependencies": {
     "@slack/client": "^3.10.0",
-    "babel-eslint": "^6.1.2",
     "babel-plugin-transform-runtime": "^6.15.0",
     "bcrypt": "^1.0.3",
     "body-parser": "^1.15.2",
@@ -67,13 +67,20 @@
   },
   "devDependencies": {
     "babel-cli": "^6.11.4",
-    "babel-eslint": "^6.1.2",
+    "babel-eslint": "^8.2.2",
     "babel-preset-es2015": "^6.13.2",
     "babel-preset-stage-0": "^6.5.0",
-    "eslint": "^3.3.1",
-    "eslint-config-airbnb-base": "^5.0.2",
-    "eslint-plugin-babel": "^3.3.0",
-    "eslint-plugin-import": "^1.13.0",
-    "pm2": "^2.5.0"
+    "eslint": "^4.18.1",
+    "eslint-config-prettier": "^2.9.0",
+    "eslint-config-standard": "^11.0.0",
+    "eslint-config-standard-react": "^6.0.0",
+    "eslint-plugin-import": "^2.9.0",
+    "eslint-plugin-node": "^6.0.1",
+    "eslint-plugin-prettier": "^2.6.0",
+    "eslint-plugin-promise": "^3.6.0",
+    "eslint-plugin-react": "^7.7.0",
+    "eslint-plugin-standard": "^3.0.1",
+    "pm2": "^2.5.0",
+    "prettier": "^1.11.0"
   }
 }
