@@ -1,15 +1,17 @@
-import mongoose from 'mongoose';
-import { User } from '../src/models';
-import config from 'config';
-import _ from 'lodash';
-import moment from 'moment';
-import 'moment/locale/th';
-import { readFileSync, writeFileSync } from 'fs';
-import path from 'path';
-import Papa from 'papaparse';
+import mongoose from 'mongoose'
+import { User } from '../src/models'
+import config from 'config'
+import _ from 'lodash'
+import moment from 'moment'
+import 'moment/locale/th'
+import { readFileSync, writeFileSync } from 'fs'
+import path from 'path'
+import Papa from 'papaparse'
 
-mongoose.Promise = global.Promise;
-mongoose.connect(process.env.MONGODB_URI || process.env.MONGOLAB_URI || config.MONGODB_URI);
+mongoose.Promise = global.Promise
+mongoose.connect(
+  process.env.MONGODB_URI || process.env.MONGOLAB_URI || config.MONGODB_URI,
+)
 
 // const backup = {
 //   programming: ['PG21', 'PG28', 'PG23', 'PG40', 'PG32', 'PG16', 'PG48', 'PG45'],
@@ -18,30 +20,91 @@ mongoose.connect(process.env.MONGODB_URI || process.env.MONGOLAB_URI || config.M
 //   content: ['CT49', 'CT10', 'CT33', 'CT24', 'CT45', 'CT42']
 // };
 
-const selectedField = ['firstName', 'lastName', 'nickname', 'facebook', 'birthdate', 'sex', 'phone', 'email', 'religion', 'university', 'academicYear', 'faculty', 'department', 'shirtSize', 'disease', 'food', 'foodAllergy', 'medAllergy', 'address', 'province', 'postalCode', 'blood', 'emergencyPhone', 'emergencyPhoneRelated', 'emergencyName'].join(' ');
+const selectedField = [
+  'firstName',
+  'lastName',
+  'nickname',
+  'facebook',
+  'birthdate',
+  'sex',
+  'phone',
+  'email',
+  'religion',
+  'university',
+  'academicYear',
+  'faculty',
+  'department',
+  'shirtSize',
+  'disease',
+  'food',
+  'foodAllergy',
+  'medAllergy',
+  'address',
+  'province',
+  'postalCode',
+  'blood',
+  'emergencyPhone',
+  'emergencyPhoneRelated',
+  'emergencyName',
+].join(' ')
 
-const list = readFileSync(path.join(__dirname, './backup.txt')).toString().split('\n');
+const list = readFileSync(path.join(__dirname, './backup.txt'))
+  .toString()
+  .split('\n')
 // console.log(list);
 
-Promise.all(list.map(interviewRef => (
-  User.findOne({ interviewRef })
-    .select(selectedField)
-    .lean()
-)))
-.then(users => users.map(user => ({
-  ...user,
-  name: `${user.firstName} ${user.lastName}`,
-  address: `${user.address} ${user.province} ${user.postalCode}`.split('\n').join(' '),
-  emergencyContactName: `${user.emergencyName} (${user.emergencyPhoneRelated})`,
-  facebook: `https://www.facebook.com/${user.facebook}`,
-  birthdate: moment(user.birthdate).format('D MMMM YYYY')
-})))
-.then((users) => Papa.unparse({
-  fields: ['name', 'nickname', 'facebook', 'birthdate', 'sex', 'phone', 'email', 'religion', 'university', 'academicYear', 'faculty', 'department', 'shirtSize', 'disease', 'food', 'foodAllergy', 'medAllergy', 'address', 'blood', 'emergencyContactName', 'emergencyPhone', '_id'],
-  data: users
-}))
-.then(data => writeFileSync('backup.csv', data, { encoding: 'utf-8' }))
-.then(() => console.log('done'));
+Promise.all(
+  list.map(interviewRef =>
+    User.findOne({ interviewRef })
+      .select(selectedField)
+      .lean(),
+  ),
+)
+  .then(users =>
+    users.map(user => ({
+      ...user,
+      name: `${user.firstName} ${user.lastName}`,
+      address: `${user.address} ${user.province} ${user.postalCode}`
+        .split('\n')
+        .join(' '),
+      emergencyContactName: `${user.emergencyName} (${
+        user.emergencyPhoneRelated
+      })`,
+      facebook: `https://www.facebook.com/${user.facebook}`,
+      birthdate: moment(user.birthdate).format('D MMMM YYYY'),
+    })),
+  )
+  .then(users =>
+    Papa.unparse({
+      fields: [
+        'name',
+        'nickname',
+        'facebook',
+        'birthdate',
+        'sex',
+        'phone',
+        'email',
+        'religion',
+        'university',
+        'academicYear',
+        'faculty',
+        'department',
+        'shirtSize',
+        'disease',
+        'food',
+        'foodAllergy',
+        'medAllergy',
+        'address',
+        'blood',
+        'emergencyContactName',
+        'emergencyPhone',
+        '_id',
+      ],
+      data: users,
+    }),
+  )
+  .then(data => writeFileSync('backup.csv', data, { encoding: 'utf-8' }))
+  .then(() => console.log('done'))
 
 // const queryPromise = interviewRef =>
 //   User.find({ interviewRef })
