@@ -1,13 +1,13 @@
-import http from 'http';
+import http from 'http'
 
-import app from '../app';
-import { Queue, Score } from '../models';
+import app from '../app'
+import { Queue, Score } from '../models'
 
-require('babel-polyfill');
-const PORT = process.env.PORT || 3001;
+require('babel-polyfill')
+const PORT = process.env.PORT || 3001
 
-const server = http.createServer(app);
-const io = require('socket.io')(server);
+const server = http.createServer(app)
+const io = require('socket.io')(server)
 
 // const sendCurrentQueue = () => (
 //   Promise.all([
@@ -26,21 +26,19 @@ const io = require('socket.io')(server);
 //   ))
 // );
 
-const sendCurrentScore = () => (
-  Score.find({})
-    .then((scores) => io.emit('score', scores))
-);
+const sendCurrentScore = () =>
+  Score.find({}).then(scores => io.emit('score', scores))
 
 io.on('connection', () => {
   // sendCurrentQueue();
-  sendCurrentScore();
-});
+  sendCurrentScore()
+})
 
 server.on('request', req => {
   // req.ioSendQueue = sendCurrentQueue;
-  req.ioSendScore = sendCurrentScore;
-});
+  req.ioSendScore = sendCurrentScore
+})
 
 server.listen(PORT, () => {
-  console.log(`Listening on port ${PORT}`);
-});
+  console.log(`Listening on port ${PORT}`)
+})
