@@ -1,20 +1,23 @@
-import { CronJob } from 'cron';
-import { exec } from 'child_process';
-import moment from 'moment';
-import slackUtils from './utilities/slack';
+import { CronJob } from 'cron'
+import { exec } from 'child_process'
+import moment from 'moment'
+import slackUtils from './utilities/slack'
 
 const backupDatabase = new CronJob({
   cronTime: '0 1 * * *',
   onTick: () => {
-    exec(`mongodump --db ywc15 --out backup/${moment().format('DD-MM-YYYY')}`, (err) => {
-      if (err) {
-        throw new Error(err);
-      }
-      slackUtils.cronBackupSuccess();
-    });
-  }
-});
+    exec(
+      `mongodump --db ywc15 --out backup/${moment().format('DD-MM-YYYY')}`,
+      err => {
+        if (err) {
+          throw new Error(err)
+        }
+        slackUtils.cronBackupSuccess()
+      },
+    )
+  },
+})
 
-backupDatabase.start();
+backupDatabase.start()
 
-console.log('Cron has been start');
+console.log('Cron has been start')
