@@ -1,2 +1,2 @@
-export { generateUniqueString } from './unique-string';
-export { uploader } from './uploader';
+export { generateUniqueString } from './unique-string'
+export { uploader } from './uploader'
