@@ -1,4 +1,9 @@
-export { authenticator, isAuthenticated, isInterviewMember, afterAnnounce } from './authenticator';
+export {
+  authenticator,
+  isAuthenticated,
+  isInterviewMember,
+  afterAnnounce,
+} from './authenticator'
 export {
   registerValidator,
   validator,
@@ -10,7 +15,12 @@ export {
   validateFileStep2,
   validateFileStep3,
   hasFile,
-  validateConfirm
-} from './validator';
-export { singleUpload } from './multer-uploader';
-export { requireRoles, requireMatchedMajor, permissions, adminAuthorize } from './admin';
+  validateConfirm,
+} from './validator'
+export { singleUpload } from './multer-uploader'
+export {
+  requireRoles,
+  requireMatchedMajor,
+  permissions,
+  adminAuthorize,
+} from './admin'
