@@ -1,10 +1,10 @@
-import mongoose from 'mongoose';
-import Promise from 'bluebird';
-mongoose.Promise = Promise;
+import mongoose from 'mongoose'
+import Promise from 'bluebird'
+mongoose.Promise = Promise
 
 const schema = new mongoose.Schema({
   __v: { type: Number, select: false },
-  string: String
-});
+  string: String,
+})
 
-export default mongoose.model('Unique', schema);
+export default mongoose.model('Unique', schema)
