@@ -1,12 +1,12 @@
-import { Router } from 'express';
-import bcrypt from 'bcrypt';
+import { Router } from 'express'
+import bcrypt from 'bcrypt'
 
-import { adminAuthen } from '../middlewares/authenticator';
-import { requireRoles, adminAuthorize } from '../middlewares';
-import { respondResult, respondSuccess, respondErrors } from '../utilities';
-import { Admin } from '../models';
+import { adminAuthen } from '../middlewares/authenticator'
+import { requireRoles, adminAuthorize } from '../middlewares'
+import { respondResult, respondSuccess, respondErrors } from '../utilities'
+import { Admin } from '../models'
 
-const router = Router();
+const router = Router()
 
 // router.post('/', requireRoles('SuperAdmin', 'Supporter'), async (req, res) => {
 //   try {
@@ -92,38 +92,38 @@ const router = Router();
 // });
 router.get('/', adminAuthen('admin'), async (req, res) => {
   try {
-    const adminUsers = await Admin.find();
-    return res.send(adminUsers);
+    const adminUsers = await Admin.find()
+    return res.send(adminUsers)
   } catch (err) {
-    return res.error(err);
+    return res.error(err)
   }
-});
+})
 
 router.post('/', adminAuthen('admin'), async (req, res) => {
   try {
     await Admin.create({
       username: req.body.username,
       password: bcrypt.hashSync(req.body.password, 10),
-      role: req.body.role
-    });
-    return res.send({ success: true });
+      role: req.body.role,
+    })
+    return res.send({ success: true })
   } catch (err) {
-    return res.error(err);
+    return res.error(err)
   }
-});
+})
 
 router.get('/me', adminAuthen('any'), (req, res) => {
-  res.send(req.admin);
-});
+  res.send(req.admin)
+})
 
 router.delete('/:id', adminAuthen('admin'), async (req, res) => {
   try {
-    await Admin.remove({ _id: req.params.id });
-    return res.send({ success: true });
+    await Admin.remove({ _id: req.params.id })
+    return res.send({ success: true })
   } catch (e) {
-    return res.error(e);
+    return res.error(e)
   }
-});
+})
 
 // router.get('/:id', requireRoles('SuperAdmin', 'Supporter'), async (req, res) => {
 //   try {
@@ -133,4 +133,4 @@ router.delete('/:id', adminAuthen('admin'), async (req, res) => {
 //     respondErrors(res)(err);
 //   }
 // });
-export default router;
+export default router
