@@ -1,15 +1,15 @@
-import { Router } from 'express';
-import fb from 'fb';
-import config from 'config';
+import { Router } from 'express'
+import fb from 'fb'
+import config from 'config'
 
-import { User } from '../models';
-import { confirm } from '../controllers/interview';
-import { singleUpload, validateConfirm } from '../middlewares';
-import { adminAuthen } from '../middlewares/authenticator';
+import { User } from '../models'
+import { confirm } from '../controllers/interview'
+import { singleUpload, validateConfirm } from '../middlewares'
+import { adminAuthen } from '../middlewares/authenticator'
 
-import programmingQuestion from './announcement.json';
+import programmingQuestion from './announcement.json'
 
-const router = Router();
+const router = Router()
 
 // router.get('/', async (req, res) => {
 //   try {
@@ -40,19 +40,24 @@ const router = Router();
 // });
 
 router.get('/', (req, res) => {
-  return res.send(programmingQuestion);
-});
+  return res.send(programmingQuestion)
+})
 
-router.get('/:refId', adminAuthen(['admin', 'programming', 'design', 'marketing', 'content']), async (req, res) => {
-  try {
-    const user = await User.findOne({ interviewRef: req.params.refId })
-      .populate('questions');
-    if (!user) return res.error('Not found');
-    return res.send(user);
-  } catch (e) {
-    return res.error(e);
-  }
-});
+router.get(
+  '/:refId',
+  adminAuthen(['admin', 'programming', 'design', 'marketing', 'content']),
+  async (req, res) => {
+    try {
+      const user = await User.findOne({
+        interviewRef: req.params.refId,
+      }).populate('questions')
+      if (!user) return res.error('Not found')
+      return res.send(user)
+    } catch (e) {
+      return res.error(e)
+    }
+  },
+)
 // router.post('/confirm', singleUpload('slip', 'jpg', 'png', 'jpeg'), validateConfirm, confirm);
 
-export default router;
+export default router
