@@ -1,1 +1,2 @@
-export const majorAsText = major => `Web ${major.charAt(0).toUpperCase()}${major.slice(1)}`;
+export const majorAsText = major =>
+  `Web ${major.charAt(0).toUpperCase()}${major.slice(1)}`
