@@ -1,2 +1,2 @@
-export { respondErrors, respondResult, respondSuccess } from './response';
-export { filterSelectedFields } from './fotmat-data';
+export { respondErrors, respondResult, respondSuccess } from './response'
+export { filterSelectedFields } from './fotmat-data'
